planter/requirements.txt
django
djangorestframework
markdown
django-filter
Pillow
drf-spectacular